Fixing the AudioSourceReader so that it can handle floats or doubles - replace dummyBuffers with a byte vector that we size depending on the bits per sample.

+++ modules/juce_audio_plugin_client/ARA/juce_ARAAudioReaders.h
@@ -43,7 +43,7 @@ private:
     // When readSamples is not reading all channels,
     // we still need to provide pointers to all channels to the ARA read call.
     // So we'll read the other channels into this dummy buffer.
-    std::vector<float> dummyBuffer;
+    std::vector<uint8_t> dummyBuffer;
 };
 
 //==============================================================================
